requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-secret-manager==2.23.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
requests==2.32.3
rsa==4.9
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
vonage==4.4.0
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.1
vonage-jwt==1.1.5
vonage-messages==1.4.0
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.6
vonage-numbers==1.0.4
vonage-sms==1.1.5
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.2.0
vonage-voice==1.3.0
Werkzeug==3.1.3
zipp==3.21.0